backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
bcrypt==4.1.2
PyJWT==2.8.0
python-socketio==5.10.0